+++ Models/Ventas.cs
@@ -10,8 +10,8 @@ namespace Alvarado_P2_AP2.Models{
         public int VentaId { get; set; }
         public DateTime Fecha { get; set; }
         public int ClienteId { get; set; }
-        public double Monto { get; set; }
-        public double Balance { get; set; }
+        public float Monto { get; set; }
+        public float Balance { get; set; }
 
         public virtual Clientes Cliente { get; set; }
